will this break it

+++ DataModules/inc/Mppt.hpp
@@ -26,8 +26,8 @@ namespace SolarGators::DataModules
 	protected:
 		//float inputVoltage;
 		//float inputCurrent;
-		uint32_t inputVoltage;
-		uint32_t inputCurrent;
+		float inputVoltage;
+		float inputCurrent;
 	};
 
 	class Mpptx1 final: public DataModule {
